old/requirements.txt
Flask
boto3
gunicorn
python-dotenv
beautifulsoup4
spacy
mmh3
flask-cors
psycopg2-binary
cassandra-driver
cassandra-sigv4